+++ library/include/gaussian_noise.hpp
@@ -6,7 +6,7 @@
 namespace cps {
     class GaussianNoise : public Signal {
     public:
-        GaussianNoise(double amplitude, int initialTimeSec, int durationSec);
+        GaussianNoise(double amplitude, double initialTimeSec, double durationSec);
 
     public:
         double value(double /* x */) const override;
